Set up CI with Azure Pipelines
On the Review your pipeline YAML, click Save and run to save the pipeline and queue a new build.

On the Save and run pane, accept the default settings and click Save and run.

Note: For the purposes of this lab, you can commit this new file directly to the master branch.

Note: It will take a moment for the pipeline to complete. During this time it will configure the build agent, pull in the source from GitHub, and build it according to the pipeline definition.

On the Summary tab of the build job's pane, verify that the build completed successfully. [skip ci]